Added PHPUnit tests for all Auth filters
